+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## - 2026-02-24
+### Added
+- SplineGuideを追加しました。
+- RouteManagerを追加しました。分岐の制御などが可能になります。
+### Changed
+- `SplineAdvanceSystem.CalcSpline`で、ログが大量発生するバグの修正
+- `SplineAdvanceSystem.GetOffsetOnSpline`のバグ修正
+- `SplineAdvanceInstantiate`を外部参照不可化
+
 ## - 2026-01-24
 ### Added
 - 指定された点から勾配を取得するAPIを追加しました。